% Generated by roxygen2 (4.0.2): do not edit by hand
\name{REACTOMEEnrichmentAnalysis}
\alias{REACTOMEEnrichmentAnalysis}
\title{REACTOME Enrichment Analysis}
\usage{
REACTOMEEnrichmentAnalysis(identifiers, resource = "UNIPROT",
  projection = FALSE, getToken = FALSE, token = FALSE)
}
\arguments{
\item{identifiers}{The identifiers of biological entities for which an enrichment analysis
is requested.}

\item{resource}{The name of the resource to which the provided identifiers belong to. Can be any of
"UNIPROT", "TOTAL", "ENSEMBL", "CHEBI", "NCBI_PROTEIN", "EMBL" or "COMPOUND". Defaults to "UNIPROT".}

\item{projection}{Boolean, \code{TRUE} implies that orthology tables in REACTOME should be used to map the given identifiers
from their original species to human. This option allows to make better use of the resource as human pathways receive a lot more
attention in Reactome. Defaults to \code{FALSE}}

\item{getToken}{Boolean, \code{TRUE} implies the token for the analysis should be returned instead of a data.frame containing
the results. Through the token and direct REST calls, the user can obtain more data about the enrichment analysis.}

\item{token}{A token obtained from a previously done analysis (tokens can expire after 7 days). Using a token will ignore
any identifiers given.}
}
\value{
a data.frame with the result of the enrichment analysis. If "getToken" is set to true, the token of the
analysis is retrieved instead.
}
\description{
Sends a request to REACTOME to perform a simple enrichment analysis with the provided
identifiers.
}

